requirements.txt
zai-sdk>=0.2.2
httpx>=0.27.0
gradio>=4.0.0
python-dotenv>=1.0.0
